similar to vCalendar, FullCalendar uses exclusive end dates
...in practice, this means date + 1 day (that is: it ends on that day at
midnight, so don't count the day it ends)

Of course, this is different from what it used to be, but it is more
like existing calendar data.

It's better to change this in the JSON than in the JS, as it's
precompiled instead of having to walk all the JS each time on the
client.