Fix the twitter link in author bio
The regexp condition was too lose, and also applied on markdown link.
So I added a few safeguard, and also fixed a corner case with string
where the link is the last word on the line such as this:

  foo http://example.com/somepage

who was converted to:

  foo <http://example.com/somepag>e

I am not sure on where this is used, but it work fine for my limited
testing.